backend/requirements.txt
fastapi
uvicorn
gitpython
langchain
chromadb
google-generativeai
python-dotenv
tiktoken
